requirements.txt
Flask==3.0.0
google-generativeai==0.3.2
gspread==6.0.0
oauth2client==4.1.3
python-dotenv==1.0.0